+++ sample-notificator-app/src/main/groovy/life/qbic/samplenotificator/datasource/notification/create/FetchSubscriberDbConnector.groovy
@@ -61,7 +61,6 @@ class FetchSubscriberDbConnector implements FetchSubscriberDataSource{
         return foundNotifications
     }
 
-
     @Override
     List<Subscriber> getSubscriberForProject(String projectCode){