+++ Changelog.rst
@@ -9,9 +9,15 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
+* Possibility to generate and send HTML based emails informing subscriber about changes in sample status in a project (`#22 <https://github.com/qbicsoftware/sample-notificator-cli/pull/22>`_)
+
+* Introduce logging, Admin notification email template and email sending if the tool failed (`#28 <https://github.com/qbicsoftware/sample-notificator-cli/pull/28>`_)
+
 **Fixed**
 
 **Dependencies**
+
+* Add ``org.jsoup:jsoup:1.13.1``
 
 **Deprecated**